rsakey

+++ trunk/code/ElectronicElections/src/RSA/RSAKey.h
@@ -13,7 +13,7 @@ typedef long long int64;
 struct RSAKey
 {
 	int64 n;
-	int64 exp;
+	int exp;
 };
